refactor(features): refactor imports to use index.ts in shared components
